feedback controller a medias

+++ src/main/java/com/workable_sb/workable/repositories/FeedbackRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.workable_sb.workable.models.Feedback;
 
-public interface FeedbackRepository extends JpaRepository <Feedback, Short>{
+public interface FeedbackRepository extends JpaRepository <Feedback, Integer>{
 
 }
